gcc -Wall lint: 	* client.c (copy_a_file): Declare p inside the #ifdef in which is 	it used. 	* commit.c (remove_file): Remove unused variable p. 	* commit.c (checkaddfile): Remove unused variables p. 	* rcs.c (RCS_isbranch): Remove unused variable p. 	* rcs.c: Remove unused declarations and definitions of 	parse_rcs_proc, rcsnode_delproc, rcslist, and repository. 	* rtag.c (rtag_fileproc): Remove unused variable p. 	* tag.c (val_fileproc): Remove unused variable node. 	* client.c, import.c, lock.c, server.c: Cast pid_t to long before 	passing it to %ld.
	* cvs.h: Don't prototype gethostname; merely declare it (on linux,
	second argument is size_t not int).

CVS-ID: src/ChangeLog 1.494
CVS-ID: src/client.c 1.102
CVS-ID: src/commit.c 1.53
CVS-ID: src/cvs.h 1.75
CVS-ID: src/import.c 1.48
CVS-ID: src/lock.c 1.24
CVS-ID: src/rcs.c 1.30
CVS-ID: src/rtag.c 1.30
CVS-ID: src/server.c 1.68
CVS-ID: src/tag.c 1.35

+++ src/lock.c
@@ -170,7 +170,7 @@ Reader_Lock (xrepository)
 #else
 		"%s.%ld", CVSRFL,
 #endif
-		getpid ());
+		(long) getpid ());
 
     /* remember what we're locking (for lock_cleanup) */
     repository = xrepository;
@@ -311,7 +311,7 @@ write_lock (repository)
 #else
 	    "%s.%ld", CVSWFL,
 #endif
-	getpid());
+	(long) getpid());
 
 #ifdef BOGUS_UNLESS_PROVEN_OTHERWISE
     /* make sure we can write the repository */
