+++ src/server.c
@@ -676,7 +676,8 @@ receive_file (size, file, gzipped)
     if (gzip_pid)
     {
 	if (waitpid (gzip_pid, &gzip_status, 0) != gzip_pid)
-	    error (1, errno, "waiting for gunzip process %ld", gzip_pid);
+	    error (1, errno, "waiting for gunzip process %ld",
+		   (long) gzip_pid);
 	else if (gzip_status != 0)
 	    error (1, 0, "gunzip exited %d", gzip_status);
     }
@@ -3409,7 +3410,8 @@ server_updated (file, update_dir, repository, updated, file_info, checksum)
 		if (fclose (f) == EOF)
 		    error (1, errno, "reading %s", short_pathname);
 		if (waitpid (gzip_pid, &gzip_status, 0) == -1)
-		    error (1, errno, "waiting for gzip process %ld", gzip_pid);
+		    error (1, errno, "waiting for gzip process %ld",
+			   (long) gzip_pid);
 		else if (gzip_status != 0)
 		    error (1, 0, "gzip exited %d", gzip_status);
 		/* Prepending length with "z" is flag for using gzip here.  */
@@ -4051,7 +4053,7 @@ error ENOMEM Virtual memory exhausted.\n");
 	strcat (server_temp_dir, "/cvs-serv");
 
 	p = server_temp_dir + strlen (server_temp_dir);
-	sprintf (p, "%ld", getpid ());
+	sprintf (p, "%ld", (long) getpid ());
     }
 
     (void) SIG_register (SIGHUP, server_cleanup);
